hw/tpm: Add tpm_crb_svsm
The regular tpm_crb interface implements CRB registers and
state-machine and send the TPM command it receives from the upper layers
to the registered TPM backend (e.g., swtpm).

For SVSM-vTPM, we can implement a new backend. However, the commands
and data from the TPM would pass through the hypervisor and host OS.

Add a new crb interface that contains the boilerplate code such that the
entirety of CRB (registers and statemachine) is handled in SVSM.